PMT #113305: Fix search bug
When searching for a string of words that mostly matched a known title,
the expected result would not appear. This was caused by a hack used to
to get Lunr to handle fields as facets, which it can not do. I convinced
myself of this fact after several hours of reading, and re-reading the
Lunr docs.

Rather, the fix filters out items that don't have the specified tag. The
tests have been updated to specifically test for the bug in this PMT.