PMT #114394: Fix search in Safari
LunrJS has some undefined behavoir in Safari, documented here:
https://github.com/olivernn/lunr.js/issues/279

This workaround console-logs out the terms which seems to slow down things
enough for it to work, suggesting a race condition somewhere. As this is
a bug with either Safari or LunrJS, we'll have to wait for them to sort
it out.